Batch input
